simplified trigger logic
